requirements.txt
python-dotenv
pytz
pywin32
tkinterdnd2